feat(inputDetailsView): Created separate data for the tests, made test pass with null data for mocked failure json
