+++ src/Analysis/GWASResults/Utils/IsJsonString.js
@@ -1,0 +1,10 @@
+const isJsonString = (str) => {
+  try {
+    JSON.parse(str);
+  } catch (e) {
+    return false;
+  }
+  return true;
+};
+
+export default isJsonString;
